Split up podcast-related repositories
This prepares the repositories for some generalization
